refactoring, added undersore terminus mode
Refactored most of the completion set manipulations, such as gathering,
organizing and serializing to its own class. This was done because I
will be repeating the same thing again for underscore terminus mode, and
copying and pasting is bad.

Added underscore terminus mode, which biases the completions offered
to words that end with an underscore when the input prefix ends with
an underscore.