generalize use of internal std::string to String
This way I can change between differing string implementations. This is
part of an ongoing experiment to minimize memory usage, as abbreviations
are usually just a couple of letters.

For now, I have switched to ustring, a small, read-only custom string
implementation. This saves around 20MB of memory based on a test with
the OGRE tags file so far.

Eventually, I will probably improve this with a custom allocator with
does memory pooling, but I am happy with memory savings for now.