rough first pass implementation of async completion
Full credit to maralla/completor.vim, which this pattern is completely
based from.

This decision is done due to an attempt to eliminate lag and flickering
from typing too fast in a file with lots of syntax highlighting. The
previous behavior is synchronous and would try to popup after every
keypress.